Fixes xml:base and default namespace declarations
The fact that these have presumably been left at initial (Protégé-suggested?) default settings happens to be inconsequential in the current version of the document, because all IRIs are given as absolute ones, and all XML elements use an explicit namespace, including the ones from this ontology.

That said, arguably they should be set to meaningful values, so that if in the future an ontology editor were to use a relative URI or a term from this ontology without namespace, then the presumably "correct  thing" would happen. For example, setting the default namespace (`xmlns`) correctly means you can write
```xml
<NOMEN_0000001>1960</NOMEN_0000001>
```
instead of
```
<obo:NOMEN_0000001>1960</obo:NOMEN_0000001>
```
and they would both result in the exact same axiom.

The other advantage is this change would eliminate a source of confusion as to what the namespace of this ontology is, see for example pensoft/OpenBiodiv#44.